+++ R/coercion.r
@@ -319,7 +319,7 @@ setMethod("as.interval", signature("logical"), function(x, start, ...) {
 #' @examples
 #' span <- interval(ymd_hms("2009-01-01 00:00:00"), ymd_hms("2010-02-02 01:01:01")) #interval
 #' as.period(span)
-#' as.period(span, units = "day")
+#' as.period(span, unit = "day")
 #' "397d 1H 1M 1S"
 #' leap <- interval(ymd("2016-01-01"), ymd("2017-01-01"))
 #' as.period(leap, unit = "days")
@@ -350,7 +350,9 @@ setMethod("as.period", signature(x = "numeric"), function(x, unit = "second", ..
   x <- as.numeric(x)
   if (missing(unit)) unit <- "second"
   unit <- standardise_date_names(unit)
-  f <- match.fun(paste(unit, "s", sep = ""))
+  f <- get(paste(unit, "s", sep = ""),
+           envir = asNamespace("lubridate"),
+           mode = "function", inherits = FALSE)
   f(x)
 })
 
@@ -626,20 +628,40 @@ setMethod("as.character", signature(x = "Interval"), function(x, ...) {
 #' as_date(10)
 #' @export
 setGeneric(name = "as_date",
-           def = function(x, ...) standardGeneric("as_date"),
-           useAsDefault = as.Date)
+           def = function(x, ...) standardGeneric("as_date"))
+
+#' @rdname as_date
+#' @export
+setMethod("as_date", "ANY",
+          function(x, ...) {
+            ## From: Kurt Hornik <Kurt.Hornik@wu.ac.at>
+            ## Date: Tue, 3 Apr 2018 18:53:19
+            ##
+            ## `zoo` has its own as.Date for which it registers its yearmon
+            ## method (and base::as.Date as the default S3 method).  In fact,
+            ## zoo also exports as.Date.yearmon etc, but the above
+            ##
+            ##    lubridate::as_date(zoo::as.yearmon("2011-01-07"))
+            ##
+            ## does not attach the zoo exports, hence does not find
+            ## as.Date.yearmon on the search path.
+            if (inherits(x, c("yearmon", "yearqtr")))
+              zoo::as.Date(x, ...)
+            else
+              base::as.Date(x, ...)
+          })
 
 #' @rdname as_date
 #' @export
 setMethod(f = "as_date", signature = "POSIXt",
-          function (x, tz = NULL) {
+          function(x, tz = NULL) {
             tz <- if (is.null(tz)) tz(x) else tz
             as.Date(x, tz = tz)
           })
 
 #' @rdname as_date
 setMethod(f = "as_date", signature = "numeric",
-          function (x, origin = lubridate::origin) {
+          function(x, origin = lubridate::origin) {
             as.Date(x, origin = origin)
           })