+++ R/durations.r
@@ -37,6 +37,7 @@ setClass("Duration", contains = c("Timespan", "numeric"), validity = check_durat
 #'   Compare,difftime,Duration-method Compare,ANY,Duration-method
 #'   Compare,Duration,Period-method Compare,Duration,difftime-method
 #'   Compare,character,Duration-method Compare,Duration,character-method
+#'   Compare,numeric,Duration-method Compare,Duration,numeric-method
 #'   as.numeric,Duration-method show,Duration-method c,Duration-method
 #'   rep,Duration-method [,Duration-method [<-,Duration,ANY,ANY,ANY-method
 #'   [[,Duration-method [[<-,Duration,ANY,ANY,ANY-method $,Duration-method
@@ -349,6 +350,18 @@ setMethod("Compare", c(e1 = "ANY", e2 = "Duration"),
           })
 
 #' @export
+setMethod("Compare", signature(e1 = "Duration", e2 = "numeric"),
+          function(e1, e2) {
+            callGeneric(e1@.Data, e2)
+          })
+
+#' @export
+setMethod("Compare", signature(e1 = "numeric", e2 = "Duration"),
+          function(e1, e2) {
+            callGeneric(e1, e2@.Data)
+          })
+
+#' @export
 setMethod("Compare", signature(e1 = "Duration", e2 = "character"),
           function(e1, e2) {
             callGeneric(e1, as.duration(e2))
@@ -365,7 +378,6 @@ setMethod("Compare", c(e1 = "difftime", e2 = "Duration"),
           function(e1, e2) {
             callGeneric(as.numeric(e1, "secs"), e2@.Data)
           })
-
 
 #' @export
 setMethod("Compare", c(e1 = "Duration", e2 = "difftime"),