+++ src/update.cpp
@@ -3,7 +3,6 @@
 #include <unordered_map>
 #include "civil_time.h"
 #include "time_zone.h"
-#include "time_zone_if.h"
 #include "utils.h"
 #include <Rcpp.h>
 
@@ -95,7 +94,7 @@ const char* get_system_tz() {
 
 const char* local_tz() {
   // initialize once per session
-  static const char* SYS_TZ = get_system_tz();
+  static const char* SYS_TZ = strdup(get_system_tz());
   const char* tz_env = std::getenv("TZ");
   if (tz_env == NULL) {
     return SYS_TZ;
@@ -235,7 +234,7 @@ Rcpp::newDatetimeVector C_update_dt(const Rcpp::NumericVector& dt,
 
   std::string tzfrom = tz_from_tzone_attr(dt);
   cctz::time_zone tzone1;
-  load_tz_or_fail(tzfrom, tzone1, "Invalid timezone of input vector: \"%s\"");
+  load_tz_or_fail(tzfrom, tzone1, "CCTZ: Invalid timezone of the input vector: \"%s\"");
 
   std::string tzto;
   cctz::time_zone tzone2;